Добавлены новые методы в InMemoryHistoryManager согласно ТЗ Отредактирован Класс Main согласно ТЗ Отредактированны Методы в InMemoryTaskManage Добавлен новый класс Node согласно ТЗ Добавлены новые тесты
