+++ src/tracker/manager/InMemoryHistoryManager.java
@@ -75,4 +75,16 @@ public class InMemoryHistoryManager implements HistoryManager {
             }
         }
     }
+    private static class Node<Task> {
+
+        public Task data;
+        public Node<Task> next;
+        public Node<Task> prev;
+
+        public Node(Node<Task> prev, Task data, Node<Task> next) {
+            this.data = data;
+            this.next = next;
+            this.prev = prev;
+        }
+    }
 }
