--- src/tracker/manager/Node.java
@@ -1,14 +1,0 @@
-package tracker.manager;
-
-public class Node<Task> {
-
-    public Task data;
-    public Node<Task> next;
-    public Node<Task> prev;
-
-    public Node(Node<Task> prev, Task data, Node<Task> next) {
-        this.data = data;
-        this.next = next;
-        this.prev = prev;
-    }
-}
